requirements.txt
Flask==2.0.1
python-dotenv==1.0.0
google-generativeai==0.3.0
asgiref==3.7.2
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
pytest==7.4.0
pytest-asyncio==0.21.1
